+++ lib/src/export.dart
@@ -1,6 +1,7 @@
 export './chart_controller.dart';
 export './chart_event.dart';
 export './chart_menu.dart';
+export './chart_overview.dart';
 export './comparison.dart';
 export './chart_style.dart';
 export './chart_translations.dart';